Remove data block boxing
Make VecDataBlock the only type returned by N5Reader to remove the need
for a boxed trait object and to allow zero-copy transformation into a
Vec (trait objects cannot be `Into`). There is no clear need for the
other conceivable backing structs to be returned (e.g., fixed array or
small vec) at this time, so this is reasonable.

To allow for unknown types to be used as blocks when writing, however,
change N5Writer to be generic over a DataBlock type reference rather
than taking a boxed trait object.

Downsides:
- An additional bound on reading methods that `VecDataBlock<T>:
  DataBlock<T>`.

Closes aschampion/rust-n5#5