Create macro for matching data types
In downstream crates it is a common problem to have to match on the N5
DataType from the DatasetAttributes and generically execute a block
dependent on the corresponding primitive type. Create a macro to do this
by creating an fn generic over the `RsType` type parameter.

Also move data type into submodule.