Use SmallVec for coordinates
Use type aliases for SmallVec for grid and block coordinates and sizes.
This reduces allocations and increases data locality. In practice,
performance differences are neglible or even negative, given the
dominance of IO, but this breaking change is made sooner rather than
later.

The stack-to-heap switch size for these SmallVecs is 6 for now, which
avoids wasting too much stack space, while covering keeping most common
N5 datasets on the stack (5D or fewer).

Reading blocks still requires an allocation for the data vec.

Closes aschampion/rust-n5#18.