Allow reading block into an existing buffer
This removes an allocation in cases where it would be a bottleneck. Note
that in the typical case allocation in insignificant relative to storage
IO. An allocation still remains for the compression trait object.

Also factors header reading. In benchmarks this is less than 1%
performance difference, and a small binary size reduction.